fix : api 1회 호출을 위해서 부모 컴포넌트에서 호출
