+++ src/components/Dashboard/index.tsx
@@ -1,13 +1,24 @@
+import { useEffect, useState } from "react";
+import { getTotal, TotalType } from "../../lib/apis";
 import Chat from "./Chat";
 import Recent from "./Recent";
 import Total from "./Total";
 
 const Dashboard = () => {
+  const [total, setTotal] = useState<TotalType[]>();
+
+  useEffect(() => {
+    const fetchTotal = async () => {
+      const data = await getTotal("이복균");
+      setTotal(data[0]);
+    };
+    fetchTotal();
+  }, []);
   return (
     <div className="w-full grid grid-cols-[2fr_1fr] gap-6">
       <div className="grid grid-cols-2 gap-6">
-        <Total mode="EARN" />
-        <Total mode="WIN" />
+        <Total mode="EARN" total={total} />
+        <Total mode="WIN" total={total} />
       </div>
       <div className="grid grid-rows-2 gap-6">
         <Recent />
